Use the buffer instead of timer to detect touches
This commit allow a more robust background detection because Chrome and Firefox doesn't throttle timer of the background tasks and that causes timing issue with the both the timer in the detector and the timer to debounce in main.js. Using a buffer structure to reduce the coupling of two timers.